Move arm_client_cache to internal/identity
