+++ v2/internal/identity/arm_client_cache_test.go
@@ -3,7 +3,7 @@
  * Licensed under the MIT license.
  */
 
-package arm
+package identity
 
 import (
 	"context"
