+++ v2/internal/reconcilers/arm/arm_client_factory.go
@@ -8,7 +8,8 @@ package arm
 import (
 	"context"
 
+	"github.com/Azure/azure-service-operator/v2/internal/identity"
 	"github.com/Azure/azure-service-operator/v2/pkg/genruntime"
 )
 
-type ARMClientFactory func(context.Context, genruntime.ARMMetaObject) (*Connection, error)
+type ARMClientFactory func(context.Context, genruntime.ARMMetaObject) (*identity.Connection, error)
