+++ v2/internal/reconcilers/arm/azure_generic_arm_reconciler_instance.go
@@ -22,6 +22,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/conversion"
 
 	"github.com/Azure/azure-service-operator/v2/internal/genericarmclient"
+	"github.com/Azure/azure-service-operator/v2/internal/identity"
 	. "github.com/Azure/azure-service-operator/v2/internal/logging"
 	"github.com/Azure/azure-service-operator/v2/internal/reconcilers"
 	"github.com/Azure/azure-service-operator/v2/internal/reflecthelpers"
@@ -39,14 +40,14 @@ type azureDeploymentReconcilerInstance struct {
 	Log              logr.Logger
 	Recorder         record.EventRecorder
 	Extension        genruntime.ResourceExtension
-	ARMClientDetails *Connection
+	ARMClientDetails *identity.Connection
 }
 
 func newAzureDeploymentReconcilerInstance(
 	metaObj genruntime.ARMMetaObject,
 	log logr.Logger,
 	recorder record.EventRecorder,
-	clientDetails *Connection,
+	clientDetails *identity.Connection,
 	reconciler AzureDeploymentReconciler) *azureDeploymentReconcilerInstance {
 
 	return &azureDeploymentReconcilerInstance{
